+++ crates/matrix-sdk-ui/src/timeline/mod.rs
@@ -59,7 +59,7 @@ use crate::timeline::pinned_events_loader::PinnedEventsRoom;
 
 mod builder;
 mod controller;
-mod day_dividers;
+mod date_dividers;
 mod error;
 mod event_handler;
 mod event_item;
@@ -175,6 +175,14 @@ impl TimelineFocus {
             TimelineFocus::PinnedEvents { .. } => "pinned-events".to_owned(),
         }
     }
+}
+
+/// Changes how dividers get inserted, either in between each day or in between
+/// each month
+#[derive(Debug, Clone)]
+pub enum DateDividerMode {
+    Daily,
+    Monthly,
 }
 
 impl Timeline {
@@ -524,6 +532,12 @@ impl Timeline {
     /// If the encryption feature is enabled, this method will transparently
     /// encrypt the room message if the room is encrypted.
     ///
+    /// The attachment and its optional thumbnail are stored in the media cache
+    /// and can be retrieved at any time, by calling
+    /// [`Media::get_media_content()`] with the `MediaSource` that can be found
+    /// in the corresponding `TimelineEventItem`, and using a
+    /// `MediaFormat::File`.
+    ///
     /// # Arguments
     ///
     /// * `path` - The path of the file to be sent.
@@ -532,6 +546,8 @@ impl Timeline {
     ///
     /// * `config` - An attachment configuration object containing details about
     ///   the attachment like a thumbnail, its size, duration etc.
+    ///
+    /// [`Media::get_media_content()`]: matrix_sdk::Media::get_media_content
     #[instrument(skip_all)]
     pub fn send_attachment(
         &self,
@@ -833,6 +849,7 @@ struct TimelineDropHandle {
     room_update_join_handle: JoinHandle<()>,
     pinned_events_join_handle: Option<JoinHandle<()>>,
     room_key_from_backups_join_handle: JoinHandle<()>,
+    room_keys_received_join_handle: JoinHandle<()>,
     room_key_backup_enabled_join_handle: JoinHandle<()>,
     local_echo_listener_handle: JoinHandle<()>,
     _event_cache_drop_handle: Arc<EventCacheDropHandles>,
@@ -853,6 +870,7 @@ impl Drop for TimelineDropHandle {
         self.room_update_join_handle.abort();
         self.room_key_from_backups_join_handle.abort();
         self.room_key_backup_enabled_join_handle.abort();
+        self.room_keys_received_join_handle.abort();
         self.encryption_changes_handle.abort();
     }
 }