+++ crates/matrix-sdk/src/event_cache/mod.rs
@@ -313,11 +313,17 @@ impl EventCache {
         prev_batch: Option<String>,
     ) -> Result<()> {
         // If the event cache's storage has been enabled, do nothing.
-        if self.inner.store.get().is_some() {
+        if self.inner.has_storage() {
             return Ok(());
         }
 
         let room_cache = self.inner.for_room(room_id).await?;
+
+        // If the linked chunked already has at least one event, ignore this request, as
+        // it should happen at most once per room.
+        if !room_cache.inner.state.read().await.events().is_empty() {
+            return Ok(());
+        }
 
         // We could have received events during a previous sync; remove them all, since
         // we can't know where to insert the "initial events" with respect to
@@ -344,6 +350,13 @@ struct AllEventsCache {
     /// A cache of related event ids for an event id. The key is the original
     /// event id and the value a list of event ids related to it.
     relations: RelationsMap,
+}
+
+impl AllEventsCache {
+    fn clear(&mut self) {
+        self.events.clear();
+        self.relations.clear();
+    }
 }
 
 struct EventCacheInner {
@@ -387,6 +400,11 @@ impl EventCacheInner {
         self.client.get().ok_or(EventCacheError::ClientDropped)
     }
 
+    /// Has persistent storage been enabled for the event cache?
+    fn has_storage(&self) -> bool {
+        self.store.get().is_some()
+    }
+
     /// Clears all the room's data.
     async fn clear_all_rooms(&self) -> Result<()> {
         // Note: one must NOT clear the `by_room` map, because if something subscribed
@@ -419,7 +437,9 @@ impl EventCacheInner {
         for (room_id, left_room_update) in updates.leave {
             let room = self.for_room(&room_id).await?;
 
-            if let Err(err) = room.inner.handle_left_room_update(left_room_update).await {
+            if let Err(err) =
+                room.inner.handle_left_room_update(self.has_storage(), left_room_update).await
+            {
                 // Non-fatal error, try to continue to the next room.
                 error!("handling left room update: {err}");
             }
@@ -429,7 +449,9 @@ impl EventCacheInner {
         for (room_id, joined_room_update) in updates.join {
             let room = self.for_room(&room_id).await?;
 
-            if let Err(err) = room.inner.handle_joined_room_update(joined_room_update).await {
+            if let Err(err) =
+                room.inner.handle_joined_room_update(self.has_storage(), joined_room_update).await
+            {
                 // Non-fatal error, try to continue to the next room.
                 error!("handling joined room update: {err}");
             }
@@ -604,7 +626,10 @@ mod tests {
 
         room_event_cache
             .inner
-            .handle_joined_room_update(JoinedRoomUpdate { account_data, ..Default::default() })
+            .handle_joined_room_update(
+                event_cache.inner.has_storage(),
+                JoinedRoomUpdate { account_data, ..Default::default() },
+            )
             .await
             .unwrap();
 
@@ -718,5 +743,29 @@ mod tests {
         // After clearing, both fail to find the event.
         assert!(room_event_cache.event(event_id).await.is_none());
         assert!(event_cache.event(event_id).await.is_none());
+    }
+
+    #[async_test]
+    async fn test_add_initial_events() {
+        // TODO: remove this test when the event cache uses its own persistent storage.
+        let client = logged_in_client(None).await;
+        let room_id = room_id!("!galette:saucisse.bzh");
+
+        let event_cache = client.event_cache();
+        event_cache.subscribe().unwrap();
+
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+        event_cache
+            .add_initial_events(room_id, vec![f.text_msg("hey").into()], None)
+            .await
+            .unwrap();
+
+        client.base_client().get_or_create_room(room_id, matrix_sdk_base::RoomState::Joined);
+        let room = client.get_room(room_id).unwrap();
+
+        let (room_event_cache, _drop_handles) = room.event_cache().await.unwrap();
+        let (initial_events, _) = room_event_cache.subscribe().await.unwrap();
+        // `add_initial_events` had an effect.
+        assert_eq!(initial_events.len(), 1);
     }
 }